+++ multipart_parser.h
@@ -34,6 +34,6 @@ struct multipart_parser_settings {
 multipart_parser* init_multipart_parser(char *boundary, const multipart_parser_settings* settings);
 void free_multipart_parser(multipart_parser* p);
 
-int multipart_parser_execute(multipart_parser* p, const char *buf, size_t len);
+size_t multipart_parser_execute(multipart_parser* p, const char *buf, size_t len);
 
 #endif
